(account) Refactor (de)serialization
